add search page, remove tips before force login.

+++ search.php
@@ -1,48 +1,28 @@
 <?php
 
-function GetRecords($user_id)
+require_once('show.php');
+
+function CheckKeyword($key_word, $record)
 {
-	$query = "select * from idpass_users where id = '$user_id'";
-	$result = mysql_query($query);
-	$row = mysql_fetch_array($result);
-	if(!is_array($row))
-	{
-		return false;
-	}
-	$user_name = $row['username'];
-	$records = array();
+	//检查记录是否包含关键字
+	if((strstr($record[0], $key_word)))
+		return true;
 	
-	//查询所有记录
-	$query = sprintf("select distinct record from idpass_secret where user_id = %d", $user_id);
-	$records_result = mysql_query($query);
-	$records_row = mysql_fetch_array($records_result);
-	if(is_array($records_row))
+	$lines = (count($record) - 1) / 3;
+	for($i = 1; $i <= $lines; $i++)
 	{
-		do
+		if($record[$i*3] == 0)
 		{
-			$record_name = $records_row[0];
-			$one_record = array($record_name);
-				
-			$query = sprintf("select name, value, encrypt from idpass_secret where user_id = %d and record = '%s'", $user_id, $record_name);
-			$result = mysql_query($query);
-			$row = mysql_fetch_array($result);
-			if(is_array($row))
-			{
-				do
-				{
-					array_push($one_record, $row[0], $row[1], $row[2]);
-				} while($row = mysql_fetch_array($result));
-			}
-			array_push($records, $one_record);
-		} while($records_row = mysql_fetch_array($records_result));
+			if(strstr($record[$i*3-2], $key_word) || strstr($record[$i*3-1], $key_word))
+				return true;
+		}
 	}
 	
-	return $records;
+	return false;
 }
 
-function ShowRecords($user_id)
+function Search($user_id, $key_word)
 {
-	//显示所有记录
 	$records = GetRecords($user_id);
 	if(!$records)
 	{
@@ -50,6 +30,7 @@ function ShowRecords($user_id)
 		return false;
 	}
 	
+	echo '<h1>' . htmlentities($key_word) . '的搜索结果：</h1><br>';
 	echo <<<STR
 <script>
 function DecryptValue(val)
@@ -59,8 +40,8 @@ function DecryptValue(val)
 		location.href='login.php';
 		return;
 	}
-	var key = CryptoJS.enc.Utf8.parse(sessionStorage.getItem('aes_key')); 
-	var iv  = CryptoJS.enc.Utf8.parse('1234567812345678'); 
+	var key = CryptoJS.enc.Utf8.parse(sessionStorage.getItem('aes_key'));
+	var iv  = CryptoJS.enc.Utf8.parse('1234567812345678');
 	return CryptoJS.AES.decrypt(val.toString(), key, { iv: iv, mode: CryptoJS.mode.CBC, padding: CryptoJS.pad.Pkcs7 }).toString(CryptoJS.enc.Utf8);;
 }
 $(document).ready(function(){
@@ -75,6 +56,8 @@ STR;
 	
 	foreach($records as $record)
 	{
+		if(!CheckKeyword($key_word, $record))
+			continue;
 		$txt .= '<li><div class="link"><label>' . $record[0] . '</label><a href="?type=deleterecord&name=' . $record[0] . '" name="delete_record">删除</a><a href="index.php?type=edit&name=' . urlencode($record[0]) . '">编辑</a></div>';
 		$txt .= '<ul class="submenu"><table>';
 		$lines = (count($record) - 1) / 3;
@@ -106,7 +89,6 @@ STR;
 	
 </script>
 STR;
-
 }
 
 ?>
